All the errors are warning. Introduce the advanced trigger which will use the photon+wire info. Make Clustering and AdvancedClustering inherit from the abstract ClusteringBase so that it's handling all the memory stuff (delete) and clustering is really only about clustering. Create the Output manager, which manages all the saving to trees. Modified the efficiency calculation.

+++ clustering/ClusterEngine.cc
@@ -175,7 +175,7 @@ void ClusterEngine::ClusterOpticalHits(std::vector<OpticalHit*>& vec_OptHit,
                   bucket[vec_TempHit[ii]].push_back(vec_TempHit[j]);
           }
       }
-      int max_threshold = 0;
+      size_t max_threshold = 0;
       OpticalHit* optical_hit_idx;
       for (size_t iii=0; iii<vec_TempHit.size(); iii++){
           size_t bucket_size = bucket[vec_TempHit[iii]].size();
